+++ SitumWayfinding/Classes/Positioning/InfoBarMapViewController.swift
@@ -8,7 +8,7 @@
 import Foundation
 import UIKit
 
-class InfoBarMapViewController: UIViewController {
+class InfoBarMapViewController: SitumViewController {
     
     @IBOutlet weak var imageView: UIImageView!
     @IBOutlet weak var mainLabel: UILabel!
@@ -20,9 +20,8 @@ class InfoBarMapViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         mainLabel.translatesAutoresizingMaskIntoConstraints = false
-        topSeparator.backgroundColor = .primaryDiminished
+        customizeUIImages()
     }
     
     func setLogo(image: UIImage?) {
@@ -51,4 +50,21 @@ class InfoBarMapViewController: UIViewController {
         }
     }
     
+}
+
+extension InfoBarMapViewController {
+    func isBeingPresented(){
+        customizeUIImages()
+    }
+    
+    func customizeUIImages() {
+        topSeparator.backgroundColor = uiColorsTheme.primaryColorDimished
+        imageView.setSitumImage(name: "swf_info", tintColor: uiColorsTheme.iconsTintColor)
+        mainLabel.textColor = uiColorsTheme.textColor
+        secondaryLabel.textColor =  uiColorsTheme.textColor
+    }
+    
+    override func reloadScreenColors(){
+        customizeUIImages()
+    }
 }