+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 All non released changes should be in CHANGELOG_UNRELEASED.md file
 
 ---------
+## [0.17.1] - 2022-12-22
+### Changed
+- Added Turkish translations
+- When the floor name is available show it instead of the ground level. The floor name will be shown in the search bar and in the bottom bar when a POI is selected
+
+## [0.17.0] - 2022-12-21
+### Added:
+- Added floor change icons in the route path to indicate when a change of floor has to be made
+- Added support to show tiles downloaded offline. To manage these tiles you should check the documentation on SitumSDK
+- Updated Situm SDK dependency to 2.55.0
+
 ## [0.16.1] - 2022-11-07
 ### Changed:
 - Improvements in the calculation of time to goal and arrival expected time whe the user request a route to a point.